--- a/test/JDBC/input/BABEL-SP_COLUMN_PRIVILEGES.mix
+++ b/test/JDBC/input/BABEL-SP_COLUMN_PRIVILEGES.mix
@@ -1,4 +1,4 @@
--- sla 12000
+-- sla 10000
 -- tsql
 CREATE DATABASE db1
 GO